Develop

+++ controllers/cards.js
@@ -3,7 +3,7 @@ const Card = require('../models/card');
 module.exports.getAllCards = (req, res) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch(() => res.status(400).send({ message: 'Произошла ошибка' }));
+    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
 };
 
 module.exports.createNewCard = (req, res) => {
@@ -12,19 +12,36 @@ module.exports.createNewCard = (req, res) => {
 
   Card.create({ name, link, owner: userId })
     .then((card) => res.status(201).send({ data: card }))
-    .catch(() => res.status(400).send({ message: 'Произошла ошибка' }));
+    .catch((err) => {
+      if (err.name === 'ValidationError') {
+        res.status(400).send({ message: 'Введите имя карточки и ссылку на картинку' });
+      } else {
+        res.status(500).send({ message: 'Произошла ошибка' });
+      }
+    });
 };
 
 module.exports.deleteCardById = (req, res) => {
-  Card.findByIdAndRemove(req.params.id)
+  Card.findById(req.params.id)
+    .orFail()
     .then((card) => {
-      if (card !== null) {
-        res.send({ data: card });
+      const { owner } = card;
+      if (req.user._id === owner.toString()) {
+        Card.findByIdAndRemove(req.params.id)
+          .then(() => res.send({ message: 'Карточка удалена' }));
       } else {
-        res.status(404).send({ message: 'Нет карточки с тaким ID' });
+        res.status(403).send({ message: 'Нельзя удалять чужую карточку' });
       }
     })
-    .catch(() => res.status(400).send({ message: 'Нет карточки с тaким ID' }));
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        return res.status(400).send({ message: 'Нет карточки с тaким ID' });
+      }
+      if (err.name === 'DocumentNotFoundError') {
+        return res.status(404).send({ message: 'Объект не найден' });
+      }
+      return res.status(500).send({ message: 'Произошла ошибка' });
+    });
 };
 
 module.exports.likeCard = (req, res) => {
@@ -41,7 +58,15 @@ module.exports.likeCard = (req, res) => {
   )
     .orFail()
     .then((card) => res.send({ data: card }))
-    .catch(() => res.status(404).send({ message: 'Что-то не так' }));
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        return res.status(400).send({ message: 'Нет карточки с тaким ID' });
+      }
+      if (err.name === 'DocumentNotFoundError') {
+        return res.status(404).send({ message: 'Объект не найден' });
+      }
+      return res.status(500).send({ message: 'Произошла ошибка' });
+    });
 };
 
 module.exports.dislikeCard = (req, res) => {
@@ -58,5 +83,13 @@ module.exports.dislikeCard = (req, res) => {
   )
     .orFail()
     .then((card) => res.send({ data: card }))
-    .catch(() => res.status(404).send({ message: 'Что-то не так' }));
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        return res.status(400).send({ message: 'Нет карточки с тaким ID' });
+      }
+      if (err.name === 'DocumentNotFoundError') {
+        return res.status(404).send({ message: 'Объект не найден' });
+      }
+      return res.status(500).send({ message: 'Произошла ошибка' });
+    });
 };
